Refactor profile page
